--User changes  -Remove Hue as a saved parameter, as well as animation parameters associated with it. It's now a GUI-only field that is never saved.  -Make histogram, density filter buffer, and all associated fields always float, even when using double. In that case, only the iteration calculations are now double. Suggested by Thomas Ludwig.  -Print all three kernels in EmberRender when the --dump_kernel option is specified.  -Apply variations filter to randoms.
--Bug fixes
 -Fix bug where hue was not being preserved when switching controllers and embers. Very hard to repro bug, but mostly overcome by eliminating hue as a saved parameter.

--Code changes
 -De-templatized DEOpenCLKernelCreator and FinalAccumOpenCLKernelCreator. They now just take a bool as a parameter to specify double precision.
 -To accommodate the buffers being float, introduce a new #define types in EmberCL called real4_bucket, and real4reals_bucket.
 -Density and spatial filtering structs now use this type.
 -ConvertDensityFilter() and ConvertSpatialFilter() no longer return a value, they just assign to the member.

+++ Source/Ember/Ember.cpp
@@ -412,6 +412,6 @@ EXPORT_TWO_TYPE_EMBER(float, float)
 
 #ifdef DO_DOUBLE
 	EXPORT_SINGLE_TYPE_EMBER(double)
-	EXPORT_TWO_TYPE_EMBER(double, double)
+	EXPORT_TWO_TYPE_EMBER(double, float)
 #endif
 }
