+++ Source/EmberAnimate/EmberAnimate.cpp
@@ -387,7 +387,7 @@ int _tmain(int argc, _TCHAR* argv[])
 #ifdef DO_DOUBLE
 		if (opt.Bits() == 64)
 		{
-			b = EmberAnimate<double, double>(opt);
+			b = EmberAnimate<double, float>(opt);
 		}
 		else
 #endif
